Move pkg/gadgets/internal/test to internal/test
This package is not useful for gadgets, so let's move it to a location
that can be imported by other packages as well.
